Refactor banking step runner into stages
