+++ src/main/scala/com/boombustgroup/amorfati/engine/economics/banking/BankCapitalSemantics.scala
@@ -79,7 +79,7 @@ object BankCapitalSemantics:
       id = "monthly-bank-capital-and-ecl-write",
       categories = Set(Category.OrdinaryPnlWaterfall, Category.EclProvisionChange),
       source = SourceAnchor(
-        "src/main/scala/com/boombustgroup/amorfati/engine/economics/banking/BankingStepRunner.scala",
+        "src/main/scala/com/boombustgroup/amorfati/engine/economics/banking/BankMultiBankStage.scala",
         "capital = capitalPnl.newCapital - eclResult.provisionChange",
       ),
       sfcTreatment = "Applies the ordinary P&L formula and the IFRS 9 provision-change hit before failure checks.",
@@ -124,8 +124,8 @@ object BankCapitalSemantics:
       id = "aggregate-exactness-capital-reconciliation",
       categories = Set(Category.ExactnessReconciliation),
       source = SourceAnchor(
-        "src/main/scala/com/boombustgroup/amorfati/engine/economics/banking/BankingStepRunner.scala",
-        "copy(capital = nextBanks(index).capital + capitalMove)",
+        "src/main/scala/com/boombustgroup/amorfati/engine/economics/banking/BankAggregateReconciliation.scala",
+        "nextBanks(index) = nextBanks(index).copy(capital = nextBanks(index).capital + capitalMove)",
       ),
       sfcTreatment = "Documented exactness patch after per-bank allocation; this is the only named residual capital writer.",
       absorber =
